Make GICv2 spurious interrupt test platform-agnostic
This test was pinned to Juno and was not included for other
platforms. In reality, it only relies on a GICv2 to be present but
does nothing specific to Juno beyond that.

Instead, we now dynamically detect the GIC version inside the test
function and skip it if it is not GICv2. The test is now part of the
generic TSP testsuite.

With the test being available to all platforms, the Juno tests XML
file is now empty, as it only contained this one test. Therefore, we
remove it.

Change-Id: I8b8f0eac905cfff8a87a2305cb2abf4bb9353ef1
Signed-off-by: Sandrine Bailleux <sandrine.bailleux@arm.com>