Update patch set 2
Patch Set 2: Code-Review-1

I think the README file of the EL3 payload should be updated to mention that the user needs to specify the CROSS_COMPILE variable now.

The build system and documentation of the EL3 payload are mostly independent of the rest of the repo (there's merely a target in the top-level Makefile that recursively calls the EL3 Makefile) so I think it's important to refer to the EL3 payload's build system requirements in its README file.

Patch-set: 2