for_each_cpu macro made more readable
The three parts of the "for" loop are more readable than before.  The
way it's now written, makes it clear that the loop starts from the
second CPU and ends whenever we reach the last one. When we have reached
the last one, tftf_topology_next_cpu returns PWR_DOMAIN_INIT and that's
how the loop is terminated.

Change-Id: I18831237840e9cfa738a48dbe7f1ec449c89f7af
Signed-off-by: John Tsichritzis <john.tsichritzis@arm.com>