Update patch set 1
Patch Set 1:

> Could you please explain what was the exact problem you faced?

Indeed it's a bit confusing. I'll try to explain and please don't hesitate to insist if I'm not clear enough :)

The problem started when I realised that Hikey doesn't have a Generic Timer. Or (at least) I haven't managed to successfully decrypt the Hikey spec so as to find where exactly it resides - if it's there, at all... Instead, Hikey utilises SP804 as the system timer.

However, TFTF code assumes that what resides at SYS_CNT_BASE1 is the Generic Timer. As such, all the register offsets and the functions that read/write those registers target the Generic Timer. But for Hikey this is not true. Hence, despite defining SYS_CNT_BASE1 to be the base address of SP804, this -obviously- doesn't work.

For example, SP804 value register resides at offset 0x4 (for the Generic Timer this is 0x0) and also it's a 32 bit register (for the Generic Timer this is a 64 bit register). Because of these differences, the function get_current_time_ms was failing for Hikey.

So, the reason why I am proposing this change is to allow platforms to use another timer as the system timer if they don't have or they have but they don't use the Generic Timer.

The reason why this cannot be mitigated only by changing the base address macros is because the differences are not just in base addresses and offsets but there might be differences in register widths also. Moreover, some offsets are part of the arch.h file so we cannot always adjust them, like we do with base addresses.

Now, you might ask "why don't you use Juno's timers"? The answer is that Juno does something strange: it uses SP804 for interrupts but the Generic Timer for get_current_time_ms (I didn't even thoroughly understood why/how this happens...) But, as a result, interrupts were not working properly even with Juno's timer drivers.

Patch-set: 1