Update patch set 1
Patch Set 1:

> I'm not sure about the name of the file. Couldn't be call it
 > "bl1_svc.h" or similar? BL1 already implies TF.

I guess you're right on the last sentence. However, I'm reluctant to call it bl1_*svc*.h for the same reason I hesitated to put the header file in include/runtime_services/: because it's not really a service... I'm not sure how to call it, though. Perhaps we could avoid answering this question by naming it bl1.h ;)

Patch-set: 1