SPM: Makefile: Add targets to build resource description blobs
The targets cactus_dtb and ivy_dtb have been added to automatically
build the resource description blobs used by Cactus and Ivy. This is
done through a new shell script: generate_dtb.sh

Example command line:

    CROSS_COMPILE=aarch64-linux-gnu- make PLAT=fvp   \
    DEBUG=1 LOG_LEVEL=50 TESTS=spm                   \
    tftf cactus cactus_dtb ivy ivy_dtb

Also, add a disclaimer to explain that the current implementation of SPM
is a prototype that is going to undergo a lot of rework.

Change-Id: Iebc3ac28d20019b59f07f70f96fefc030f1a79ce
Signed-off-by: Antonio Nino Diaz <antonio.ninodiaz@arm.com>