Update patch set 1
Patch Set 1: Code-Review-1

I agree with Antonio that the platform ports directory hierarchy could be improved.

First of all, I too think that the board/ directory could be dropped.  However, I think this is outside of the scope of this patch and should be addressed separately.

For now, I'd like to propose the following directory hierarchy for the SGI-575 platform port:

 plat/arm/
 |- common
 |- board
 |   |- fvp
 |   |- juno
 |   |- sgi
 |   |   |- common
 |   |   |- sgi575

The code that is currently provided in plat/arm/sgi in this patch stack would move to plat/arm/board/sgi/common in my proposal.

The code that is currently provided in plat/arm/board/sgi575 in this patch stack would move to plat/arm/board/sgi/sgi575 in my proposal. 

Later on, when we remove the board directory, this is what I have in mind:

 plat/arm/
 |- common
 |- fvp
 |- juno
 |- sgi
 |   |- common
 |   |- sgi575

Patch-set: 1
Label: Code-Review=-1