cactus: Reorganize stack
Create a new section for the stacks inside of BSS. SPM zeroes the
memory in any region marked as BSS, so it isn't needed to do it here.

Note that now each region (text, rodata, data, bss) need to be aligned
to a page boundary so that they can be different regions in the resource
description.

Previously the memory to be used as stack was passed in the boot info
struct. Now this is not needed as the partition can define all regions
it wants to use.

Change-Id: If330e27b0b27dde93b6c9e63b7136d23bdc7dd9e
Signed-off-by: Antonio Nino Diaz <antonio.ninodiaz@arm.com>