Update patch set 1
Patch Set 1:

> Maybe we can keep this option for the macro I've proposed in
 > https://review.trustedfirmware.org/#/c/TF-A/tf-a-tests/+/439/ ?

I think the colors should not be controlled by a build flag in any case. Either we enable them all the time or we provide a special variant of this macro that additionally uses color codes to print the CPU ID and the tests choose which one they want to use (color or no color).

Patch-set: 1