Update patch set 1
Patch Set 1:

Also, it would be good to get a new patch on top of this series to update the documentation. We should at least mention that there's a port of TF-A Tests for Hikey960 and briefly describe the level of support (only 1 core, most of the tests disabled, and so on).

Patch-set: 1