Fix infinite reboot loop in manual tests
The 'manual' test suite enters an infinite reboot loop instead of
properly tracking its state while running the test cases. This occurs
because non-volatile memory (NVM) is indiscriminately cleared each
time the test suite invokes a reboot or shutdown after running a test
case.

The correct behaviour is to clear the NVM only after completion of
the whole test suite. This ensures that the state is not lost part-way
through.

This patch overrides the default NVM-clearing behaviour (set in
defaults.mk) so that tftf_cold_boot_main() does not initialize a new
test session between test cases.

Additionally, the patch sets the USE_NVM build variable to 1 so that
it is not necessary to manually override the default of 0 for correct
behaviour. If USE_NVM is set to 0 then these tests cannot succeed as
the test results will be stored in DRAM, which does not survive a
reboot.

Change-Id: Ifc60a12a789826d1615d455b32928c5873c6f75f
Signed-off-by: Paul Beesley <paul.beesley@arm.com>