Update patch set 2
Patch Set 2: Allow-CI+2 IP-Review+1 Code-Review+2

(1 comment)

We have the same problem with the firmware update tests (which also need to retain their state across a reboot) but in their case we handle the issue via the build scripts directly. See platform-ci/tftf_config/fvp-fwu, which uses the following build options:

 CROSS_COMPILE=aarch64-linux-gnu-
 NEW_TEST_SESSION=0
 PLAT=fvp
 TESTS=fwu
 USE_NVM=1

(same for juno-fwu config).

I think your solution is better because it solves the problem in a standalone manner, with no dependency on these build scripts, so I think we should prepare a similar patch as this one for the FWU tests.

Patch-set: 2
Label: IP-Review=+1