Set required NVM defaults for FWU test suite
The firmware update (FWU) test suite needs to retain its state across
a system reboot. This means that it depends on the USE_NVM build
variable being set to 1, and the NEW_TEST_SESSION being set to 0.
Together these variables ensure that ongoing test results are stored in
non-volatile memory (NVM), and that the NVM is not cleared between
reboots.

Currently these variables are set externally for this suite from the CI
scripts but the correct behaviour should be for this suite's makefile
to contain the correct defaults so that no external arguments are
needed.

Change-Id: I06da6f42ae72ad0485d22f8573a25e5a17505227
Signed-off-by: Paul Beesley <paul.beesley@arm.com>