Update patch set 5
Patch Set 5:

(2 comments)

This comment applies to all of the comments I didn't reply to.
The goal of the patch is to import libc from TF-A and make the smallest amount of changes possible to make it work in TFTF so that both codes end up being very similar. From my point of view modifying the code defeats the purpose of the patch.
What do you think?

Patch-set: 5
Label: Code-Review=0