Update patch set 5
Patch Set 5:

(1 comment)

> This comment applies to all of the comments I didn't reply to.
 > The goal of the patch is to import libc from TF-A and make the
 > smallest amount of changes possible to make it work in TFTF so that
 > both codes end up being very similar. From my point of view
 > modifying the code defeats the purpose of the patch.
 > What do you think?

I agree that there's some value in keeping both TF-A and TF-A Tests libcs as similar as possible. But if we can identify improvements, I see no reason why we couldn't do them in both repositories! We usually import changes from TF-A into TF-A Tests when relevant but we could also do it the other way around. Both projects can benefit from each other.

This does not have to be done in this patch, though. It can come later in separate patches. Among my comments, most are suggestions of improvements. There's only one that I consider as gating, it's [1], the current behaviour of the code does not seem acceptable to me.

[1] https://review.trustedfirmware.org/#/c/TF-A/tf-a-tests/+/615/5/lib/libc/snprintf.c@203

Patch-set: 5