Update patch set 1
Patch Set 1:

> I have added CI support for this platform and submitted the patch
 > for review. Can you please review the CI patch and merge these
 > patches?

Thanks for that, I've merged the said CI patch.

However, if I am not mistaken, there's still no support for the EL3 payload in the CI. The CI patch that I've reviewed took care of building and running the TFTF on the RD-N1-Edge FVP but I have not seen anything similar for the EL3 payload.

Could you please also take care of that?

Patch-set: 1