Update patch set 1
Patch Set 1:

> CI

Ah I see, the infamous UART stall issue... That explains why it was failing only in the CI, yes :)

Could you please just edit the commit message to explain that the issue was not in TF-A Tests but rather in our CI scripts? Just for reference, if someone ever looks back at the git history, they might wonder why all of a sudden we re-enabled these tests, with no apparent reason.

Patch-set: 1