Update patch set 3
Patch Set 3: Code-Review-1

(11 comments)

If I understand correctly, the purpose of Quark is to have the smallest possible secure partition. Instead of creating a new secure partition, could we not have built one out of Cactus or Ivy, cutting down some of their features at build time?

Ivy is already duplicating much of Cactus code and now Quark here does the same. I'm worried about the proliferation of these images. Are we going to keep duplicating this code every time we have a slightly different use case? It feels to me that we should re-architecture the secure partitions code to better share the common bits and keep the code specific to each partition to the minimal.

Patch-set: 3
Label: Code-Review=-1