Update patch set 3
Patch Set 3:

Cactus and Ivy need to map the UART, which makes them unable to test TTST. In order to not map the UART, you need to do modifications to the build system so that it is not included in the resource description, and to change the drivers so that they don't actually write anything. It just looks cleaner to have another partition.

Also, I don't see the problem with having multiple partitions. In my opinion having multiple partitions is better because we might detect problems that we wouldn't with just two partitions.

Patch-set: 3
Label: Code-Review=0