Update patch set 4
Patch Set 4:

> Cactus and Ivy need to map the UART, which makes them unable to test TTST. In order to not map the UART, you need to do modifications to the build system so that it is not included in the resource description, and to change the drivers so that they don't actually write anything. 

I am not saying that it is possible today without any modifications. No doubt we would need to do some refactoring first and the modifications you've listed are part of the items we would need to refactor in order to be in a better place.

> It just looks cleaner to have another partition.
> 
> Also, I don't see the problem with having multiple partitions. In my opinion having multiple partitions is better because we might detect problems that we wouldn't with just two partitions.

It certainly does not look cleaner to me. The problem with having multiple partitions, in the context of the current code organization, is code duplication. It is never OK to duplicate code when there are alternatives. Duplicating code is a maintenance burden, it is error-prone and not scalable.

Patch-set: 4