Update patch set 4
Patch Set 4:

> I am not saying that it is possible today without any modifications. No doubt we would need to do some refactoring first and the modifications you've listed are part of the items we would need to refactor in order to be in a better place.

The current build system is really fragile and I'd rather not modify it as part of this patch stack.

> It certainly does not look cleaner to me. The problem with having multiple partitions, in the context of the current code organization, is code duplication. It is never OK to duplicate code when there are alternatives. Duplicating code is a maintenance burden, it is error-prone and not scalable.

What's the problem with having multiple partitions for different tests? Isn't the whole point of SPM to have multiple partitions that do small things? We need to have multiple partitions to test with.

There is not a lot of code duplication. The only thing I can think of is the entrypoint assembly file, that could be made common to all partitions. Otherwise, the number of files needed per partition is really small.

Not to mention than from a conceptual point of view it doesn't make sense to have Cactus and "Cactus-NoPrint" or something like that. My approach is more aligned with the problem that SPM is trying to solve.

Patch-set: 4