Update patch set 4
Patch Set 4:

> What's the problem with having multiple partitions for different tests? Isn't the whole point of SPM to have multiple partitions that do small things? We need to have multiple partitions to test with.

Yes, SPM is designed to manage several secure partitions and thus we do need to have multiple partitions to test with. That does not necessarily mean we have to have several times the same code and that each secure partition's code has to be independent. It just means we need to have several secure partitions binaries to test with. We could derive several partitions out of a common code base if they are very similar.

We have to keep in mind that we are not in a real-world use case here. We are artificially creating secure partitions that perform some dummy actions just for the purpose of testing the SPM software stack as a whole. So to borrow your words, your approach "might be more aligned with the problem that SPM is trying to solve" in the real world, yes, as partitions would be provided by different vendors in most cases and they would fulfill different responsibilities. But I think this argument does not really apply in our case here.

> There is not a lot of code duplication. The only thing I can think of is the entrypoint assembly file, that could be made common to all partitions. Otherwise, the number of files needed per partition is really small.

That is not the feeling I get when I'm reviewing this patch. I feel like I've seen most of this code (twice) already, with some slight differences. It's not just the C/asm code I'm talking about, it is all the files involving in creating a secure partition in the TF-A Tests code base.

In fact, if I run a tool (I've used similarity-checker here) to analyze the lexical similarity percentage between the 3 partitions directories, here are the results I get:

 quark/quark.ld.S consists for 91 % of ivy/ivy.ld.S material
 quark/quark.dts consists for 80 % of ivy/ivy.dts material
 quark/quark_main.c consists for 68 % of ivy/ivy_main.c material
 quark/quark.h consists for 58 % of ivy/ivy.h material
 quark/quark.mk consists for 52 % of ivy/ivy.mk material
 quark/quark_def.h consists for 35 % of ivy/ivy_def.h material
 
 quark/quark.ld.S consists for 91 % of cactus/cactus.ld.S material
 quark/quark.dts consists for 78 % of cactus/cactus.dts material
 quark/quark_main.c consists for 68 % of cactus/cactus_main.c material
 quark/quark.h consists for 58 % of cactus/cactus.h material
 quark/quark.mk consists for 52 % of cactus/cactus.mk material
 quark/quark_def.h consists for 29 % of cactus/cactus_def.h material
 
 ivy/ivy.dts consists for 100 % of cactus/cactus.dts material
 ivy/ivy.ld.S consists for 92 % of cactus/cactus.ld.S material
 ivy/ivy_main.c consists for 71 % of cactus/cactus_main.c material
 ivy/ivy.h consists for 58 % of cactus/cactus.h material
 ivy/ivy.mk consists for 56 % of cactus/cactus.mk material
 ivy/ivy_def.h consists for 45 % of cactus/cactus_def.h material

I'm quite convinced we could do a better job.

Patch-set: 4