Update patch set 4
Patch Set 4:

> Yes, SPM is designed to manage several secure partitions and thus we do need to have multiple partitions to test with. That does not necessarily mean we have to have several times the same code and that each secure partition's code has to be independent. It just means we need to have several secure partitions binaries to test with. We could derive several partitions out of a common code base if they are very similar.

Having a partition that can be built in different ways is less maintainable than individual partitions that only do one thing. Specially if you factor in the problem of the build system. It's not just modifying the C code, you need to modify the DTS files depending on build options, and that sounds like a lot of problems for the build system.

> We have to keep in mind that we are not in a real-world use case here. We are artificially creating secure partitions that perform some dummy actions just for the purpose of testing the SPM software stack as a whole. So to borrow your words, your approach "might be more aligned with the problem that SPM is trying to solve" in the real world, yes, as partitions would be provided by different vendors in most cases and they would fulfill different responsibilities. But I think this argument does not really apply in our case here.

The whole point of this repository is to have test code, so creating artificial partitions to perform dummy actions to test makes sense to me.

> That is not the feeling I get when I'm reviewing this patch. I feel like I've seen most of this code (twice) already, with some slight differences. It's not just the C/asm code I'm talking about, it is all the files involving in creating a secure partition in the TF-A Tests code base.

If you actually take a closer look at them you will see that the only files that could be made common are the linkerscripts, quark/cactus/ivy.h, and the entrypoings. The rest just happens to be the same because all partitions need to define some things. Saying there is unnecessary repetition there is like saying that we repeat a lot of code because different platforms have almost the same code in some files.

Patch-set: 4