Update patch set 4
Patch Set 4:

> Having a partition that can be built in different ways is less maintainable
> than individual partitions that only do one thing. Specially if you factor in
> the problem of the build system. It's not just modifying the C code, you need
> to modify the DTS files depending on build options, and that sounds like a lot
> of problems for the build system.

I sense you might have misunderstood what I had in mind. I am not proposing that we have a single source file for all partitions, and that we sprinkle it with #ifdef preprocessor directives to cut out specific things for each partition. That would indeed be quite ugly.

Instead, I am proposing that we come up with a data-driven framework, whereby our different secure partitions provide a description of themselves. Things like, their names, the secure services they provide, their layout, and so on.

I've tried to quickly draft this below. I've not sorted out all details but hopefully you get the idea.

The secure partition code (Ivy, Cactus, Cactus_MM, Quark) would be something like:

 struct sp ivy_sp = {
	.name = "Ivy",
	.setup = ivy_setup,
	// ...
 };

 static void print_magic(const struct *sp_args, struct sp_ret *ret) {
	INFO("IVY: Magic: 0x%x\n", IVY_MAGIC_NUMBER);
	ret->ret0 = SPRT_SUCCESS;
 }

 static void get_magic(const struct *sp_args, struct sp_ret *ret) {
	ret->ret0 = SPRT_SUCCESS;
	ret->ret1 = IVY_MAGIC_NUMBER;
 }

 static void sleep_ms(const struct *sp_args, struct sp_ret *ret) {
	sp_sleep(sp_args[2]);
	ret->ret0 = SPRT_SUCCESS;
 }

static void ivy_setup(void) {
	console_init(PL011_UART3_BASE,
		     PL011_UART3_CLK_IN_HZ,
		     PL011_BAUDRATE);

	register_service(IVY_PRINT_MAGIC, print_magic);
	register_service(IVY_GET_MAGIC, get_magic);
	register_service(IVY_SLEEP_MS, sleep_ms);
}

And then we could factorize the SPRT communication bits in the secure partition framework, like so:

 static void sp_message_handler(struct sprt_queue_entry_message *message) {
	struct sp_ret ret;

	if (message->type == SPRT_MSG_TYPE_SERVICE_REQUEST) {
		void (*handler) (const struct *sp_args, struct sp_ret *ret) =
			get_service_handler(message->args[1]);

		if (handler == NULL) {
			NOTICE("%s: Unhandled Service ID 0x%x\n", sp->name,
			       (unsigned int)message->args[1]);
			ret.ret0 = SPRT_NOT_SUPPORTED;
			break;
		}

		handler(message->args, &ret);
	} else {
		NOTICE("%s: Unhandled Service type 0x%x\n", sp->name,
		       (unsigned int)message->type);
		ret0 = SPRT_NOT_SUPPORTED;
	}

	sprt_message_end(message, ret.ret0, ret.ret1, ret.ret2, ret.ret3);
 }

 void __dead2 sp_main(void) {
	struct sp *sp;

	sp->setup();

	/*
	 * Handle secure service requests.
	 */
	sprt_initialize_queues((void *)sp->spm_buf_base);

	while (1) {
		struct sprt_queue_entry_message message;

		/*
		 * Try to fetch a message from the blocking requests queue. If
		 * it is empty, try to fetch from the non-blocking requests
		 * queue. Repeat until both of them are empty.
		 */
		while (1) {
			int err = sprt_get_next_message(&message,
					SPRT_QUEUE_NUM_BLOCKING);
			if (err == -ENOENT) {
				err = sprt_get_next_message(&message,
						SPRT_QUEUE_NUM_NON_BLOCKING);
				if (err == -ENOENT) {
					break;
				} else {
					assert(err == 0);
					sp_message_handler(&message);
				}
			} else {
				assert(err == 0);
				sp_message_handler(&message);
			}
		}

		sprt_wait_for_messages();
	}

 }

That's one way to design it. As I said, I've not completely thought this through, there might be caveats or better ways to do it.

> The whole point of this repository is to have test code, so creating
> artificial partitions to perform dummy actions to test makes sense to me.

It does make sense to me too! I did not mean that creating artificial partitions to perform dummy actions is a bad thing, I believe this is the right testing approach. I'm just saying that this is not necessarily representative of how things will work/look like in the real world, outside of our test environment. So I think that some of the design choices we make in TF-A Tests here  do not necessarily apply for real secure partitions that implement actual use cases, and vice-versa.

> If you actually take a closer look at them you will see that the only files
> that could be made common are the linkerscripts, quark/cactus/ivy.h, and the
> entrypoings. The rest just happens to be the same because all partitions need
> to define some things. Saying there is unnecessary repetition there is like
> saying that we repeat a lot of code because different platforms have almost
> the same code in some files.

Platform code is not really comparable IMHO. Every platform has different requirements and I think it's difficult to come up with common interfaces that suit every use case so I see less opportunity to share code. Also, we want to give as much freedom as possible to platform integrators, they are the ones who know best what makes sense for their platforms.

I see more opportunity to share code for secure partitions because we own this code right now, so it is a lot easier.

Patch-set: 4