Update patch set 4
Patch Set 4:

If there is way to share code without too much hassle, we should do it as it will save us maintenance effort in the long term. Having a maintainable and extensible Test framework code is essential as the project matures and enables more functionality.

Going by the new methodology for development, we could merge this initial patch if it achieves the functionality but then agree to follow up implementation till we arrive to the state we want to.

Patch-set: 4
Reviewer: Gerrit User 1000056 <1000056@8f6f209b-db1a-4cbf-aa44-c8bc30e9bfda>
Label: Allow-CI=0