Update patch set 4
Patch Set 4:

> If there is way to share code without too much hassle, we should do
 > it as it will save us maintenance effort in the long term. Having a
 > maintainable and extensible Test framework code is essential as the
 > project matures and enables more functionality.
 > 
 > Going by the new methodology for development, we could merge this
 > initial patch if it achieves the functionality but then agree to
 > follow up implementation till we arrive to the state we want to.

I think that designing a framework as Sandrine has explained is way beyond the scope of this task, and it would probably even need a discussion in the mailing list. This is surely going to take a long time, so it is better to merge this patch stack as it is.

Patch-set: 4