Update patch set 4
Patch Set 4:

> But what comments do I remove? I still need to preserve the comment to explain why putc returns x0 in x0, or to explain why init returns 1.

I think John's suggestion was to remove all comments because people can refer to a real driver implementation to figure the usage. That said, if you just want to keep the 2 pieces of information you mention (putc leaving the character in x0 and init returning 1 because it always succeeds) that's fine by me.

Patch-set: 4