Update patch set 1
Patch Set 1: Code-Review-1

I've tried running this on the AEM 4x4 FVP with debug traces on and it prints the following messages:

 NOTICE:    Cluster #0   [4 CPUs]
 NOTICE:      CPU #0   [MPID: 0x1000000]
 NOTICE:      CPU #1   [MPID: 0x1000100]
 NOTICE:      CPU #2   [MPID: 0x1000200]
 NOTICE:      CPU #3   [MPID: 0x1000300]
 NOTICE:    Cluster #1   [4 CPUs]
 NOTICE:      CPU #8   [MPID: 0x1010000]
 NOTICE:      CPU #9   [MPID: 0x1010100]
 NOTICE:      CPU #10   [MPID: 0x1010200]
 NOTICE:      CPU #11   [MPID: 0x1010300]

As you can see, the CPU IDs are incorrect for the second cluster. Without this patch, it used to print:

 NOTICE:    Cluster #0   [4 CPUs]
 NOTICE:      CPU #0   [MPID: 0x1000000]
 NOTICE:      CPU #1   [MPID: 0x1000100]
 NOTICE:      CPU #2   [MPID: 0x1000200]
 NOTICE:      CPU #3   [MPID: 0x1000300]
 NOTICE:    Cluster #1   [4 CPUs]
 NOTICE:      CPU #4   [MPID: 0x1010000]
 NOTICE:      CPU #5   [MPID: 0x1010100]
 NOTICE:      CPU #6   [MPID: 0x1010200]
 NOTICE:      CPU #7   [MPID: 0x1010300]

So it looks like the function dump_topology() in plat/common/plat_topology.c needs to be adapted.

Patch-set: 1
Label: Code-Review=-1