Update patch set 1
Patch Set 1:

Just tested that on the AEM quad-cluster FVP [1] and I got even stranger CPU IDs:

 NOTICE:    Cluster #0   [4 CPUs]
 NOTICE:      CPU #0   [MPID: 0x0]
 NOTICE:      CPU #1   [MPID: 0x1]
 NOTICE:      CPU #2   [MPID: 0x2]
 NOTICE:      CPU #3   [MPID: 0x3]
 NOTICE:    Cluster #1   [4 CPUs]
 NOTICE:      CPU #8   [MPID: 0x100]
 NOTICE:      CPU #9   [MPID: 0x101]
 NOTICE:      CPU #10   [MPID: 0x102]
 NOTICE:      CPU #11   [MPID: 0x103]
 NOTICE:    Cluster #2   [4 CPUs]
 NOTICE:      CPU #16   [MPID: 0x200]
 NOTICE:      CPU #17   [MPID: 0x201]
 NOTICE:      CPU #18   [MPID: 0x202]
 NOTICE:      CPU #19   [MPID: 0x203]
 NOTICE:    Cluster #3   [4 CPUs]
 NOTICE:      CPU #24   [MPID: 0x300]
 NOTICE:      CPU #25   [MPID: 0x301]
 NOTICE:      CPU #26   [MPID: 0x302]
 NOTICE:      CPU #27   [MPID: 0x303]

The question is whether this is localised to dump_topology() or more generally to platform_get_core_pos(). If it is the latter, this might be a worry, I can't picture whether this will cause any problem right now. I need to spend more time thinking about it.

[1] test_groups=fvp-tspd-quad-cluster,fvp-default:fvp-tftf-fip.tftf-aemv8a.quad-debug

Patch-set: 1