Update patch set 2
Patch Set 2:

> The issue of non-contiguous cpu numbering not only exists for dump_topology but also will reflect in plat_get_core_pos() since this macro uses the MPIDR value as well as FVP_MAX_CPUS_PER_CLUSTER constant. As far as I understand, this may not cause any issues functionally but will have additional memory overhead since the tftf data-structures will have unused entries. For example: entries 4, 5, 6, 7 and 12, 13, 14, 15 will never be used. (I took FVP Base_AEM_4x4 as an example here). Please correct me if I am wrong.

OK, that matches my understanding, thanks for confirming.

The memory overhead induced by this patch is not a big concern. We typically don't worry too much about memory consumption in TF-A Tests and consider we have the whole DRAM for us, as nobody else is running in the Normal World.

I am more worried about the confusion the non-contiguous CPU IDs could create. But it might be inevitable without more invasive changes to the FVP topology code...

Patch-set: 2