Update patch set 6
Patch Set 6: Code-Review-1

(2 comments)

In addition to my inline comments, I think there is a number of improvements we could do in this code in the future.

First of all, I think it would make sense to have an AArch32 version of this test file and an AArch64 version. Right now, I find the code quite difficult to read with all the #ifdef __aarch64 preprocessor directives, when in fact the AArch32 version sums up to all test functions returning TEST_RESULT_SKIPPED.

Then I think we would gain in clarity if we introduced some helper functions, for example to make a 128-bit key from two 64-bit values, to retrieve one of the 5 keys as a 128-bit value, to write a key given its two 64-bit halves, and so on.

Let's revisit after the release.

Patch-set: 6
Label: Code-Review=-1