Update patch set 7
Patch Set 7:

> 1) "First of all, I think it would make sense to have an AArch32 version of this test file and an AArch64 version"
> 
> http://arminfo.emea.arm.com/help/topic/com.arm.doc.ddi0487e.a/DDI0487E_a_armv8_arm.pdf
> 
> "ARMv8.3-PAuth, Pointer Authentication
> ...
> This feature is supported only in AArch64 state."

I know but we still want to be able to build this file for AArch32. The code supports it right now. For example, if TFTF is built for ARCH=aarch32 right now, test_pauth_reg_access() will be equivalent to:

 test_result_t test_pauth_reg_access(void)
 {
	SKIP_TEST_IF_AARCH32();
 }

So I am suggesting we have an aarch32/test_pauth.c file that simply returns TEST_RESULT_SKIPPED for all test cases:

 test_result_t test_pauth_reg_access(void)
 {
	return TEST_RESULT_SKIPPED;
 }

And we'd have a separate aarch64/test_pauth.c file, identical to the one introduced in this patch except that we'd remove all #ifdef __arch64__ lines, which IMO hurts readability.

Patch-set: 7
Label: Code-Review=0