Update patch set 8
Patch Set 8:

> 2) "Then I think we would gain in clarity if we introduced some helper functions, for example to make a 128-bit key from two 64-bit values, to retrieve one of the 5 keys as a 128-bit value, to write a key given its two 64-bit halves, and so on."
> 
> We don't have and don't need this kind of functionality in TF-A. Could you give examples when those helper functions would be required?

Just to be clear, such helper functions are not *needed*, I just think they would be nice and would improve the readability. Also, I meant helper functions local to this test file, not necessarily exported as generic library functions.

I mentioned having a helper function to "write a key given its two 64-bit halves", this is what I am targeting:

 write_apiakeylo_el1((uint64_t)plat_key);
 write_apiakeyhi_el1((uint64_t)(plat_key >> 64));

-> could be replaced by a function call like,

 void write_key_reg(uint128_t key)

Another example I gave was to "retrieve one of the 5 keys as a 128-bit value", I had this kind of code in mind:

 plat_key = read_apiakeylo_el1() |
	((uint128_t)(read_apiakeyhi_el1()) << 64);

-> could be replaced by a function call like:

 uint128_t read_key_reg(enum key_reg)

where enum key_reg would identify which of the 5 keys we want to query. 

That sort of things.

Patch-set: 8